pyproject.toml
[tool.poetry]
name = "cyberdrop-dl-patched"
version = "5.7.1"
description = "Bulk downloader for multiple file hosts"
authors = ["Jacob B <admin@script-ware.net>"]
readme = "README.md"
repository = "https://github.com/jbsparrow/CyberDropDownloader"
documentation = "https://script-ware.gitbook.io/cyberdrop-dl/"
packages = [{ include = "cyberdrop_dl" }]
include = ["CHANGELOG.md"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
aiohttp = "^3.9.1"
aiolimiter = "^1.1.0"
inquirerpy = "^0.3.4"
platformdirs = "^4.2.2"
rich = "^13.7.0"
pyyaml = "^6.0.1"
certifi = "^2024.2.2"
browser-cookie3 = "^0.19.1"
beautifulsoup4 = "^4.12.2"
filedate = "^3.0"
aiosqlite = "0.17.0"
aiofiles = "0.8.0"
asyncpraw = "^7.7.1"
myjdapi = "^1.1.7"
mediafire = "^0.6.1"
mutagen = "^1.47.0"
pillow = "^10.2.0"
get-video-properties = "^0.1.1"
humanfriendly = "^10.0"
send2trash = "^1.8.3"
arrow = "^1.3.0"
apprise = "^1.9.0"

[tool.poetry.scripts]
cyberdrop-dl = "cyberdrop_dl.main:main"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
ruff = "^0.7.2"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "N",  # PEP8 naming conventions
]

ignore = [
    "N806", # Uppercase variables in functions
    "E501", # Suppress line-too-long, let formatter decide
]


[tool.ruff.lint.pydocstyle]
convention = "google"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
Changelog = "https://github.com/jbsparrow/CyberDropDownloader/blob/master/CHANGELOG.md"
